+++ spring_prj/src/main/java/com/c4ccup/laugh/domain/Oogiri.java
@@ -3,7 +3,7 @@ package com.c4ccup.laugh.domain;
 import java.time.LocalDateTime;
 import java.util.List;
 
-import com.c4ccup.laugh.controller.OogiriAnswerResponse;
+import com.c4ccup.laugh.controller.bean.res.OogiriAnswerResources;
 
 public class Oogiri {
 
@@ -35,7 +35,7 @@ public class Oogiri {
     /** 回答削除日時 */
     private LocalDateTime answerDeletedAt;
     /** 回答リスト */
-    private List<OogiriAnswerResponse> answers;
+    private List<OogiriAnswerResources> answers;
 
     // リアクション情報
     /** リアクションID */
@@ -59,6 +59,7 @@ public class Oogiri {
 
     /**
      * お題IDを取得します。
+     * 
      * @return お題ID
      */
     public int getThemeId() {
@@ -67,6 +68,7 @@ public class Oogiri {
 
     /**
      * お題IDを設定します。
+     * 
      * @param themeId お題ID
      */
     public void setThemeId(int themeId) {
@@ -75,6 +77,7 @@ public class Oogiri {
 
     /**
      * お題ユーザーIDを取得します。
+     * 
      * @return お題ユーザーID
      */
     public int getThemeUserId() {
@@ -83,6 +86,7 @@ public class Oogiri {
 
     /**
      * お題ユーザーIDを設定します。
+     * 
      * @param themeUserId お題ユーザーID
      */
     public void setThemeUserId(int themeUserId) {
@@ -91,6 +95,7 @@ public class Oogiri {
 
     /**
      * お題ユーザーネームを取得します。
+     * 
      * @return お題ユーザーネーム
      */
     public String getThemeUserName() {
@@ -99,6 +104,7 @@ public class Oogiri {
 
     /**
      * お題ユーザーネームを設定します。
+     * 
      * @param themeUserName お題ユーザーネーム
      */
     public void setThemeUserName(String themeUserName) {
@@ -107,6 +113,7 @@ public class Oogiri {
 
     /**
      * お題内容を取得します。
+     * 
      * @return お題内容
      */
     public String getThemeContent() {
@@ -115,6 +122,7 @@ public class Oogiri {
 
     /**
      * お題内容を設定します。
+     * 
      * @param themeContent お題内容
      */
     public void setThemeContent(String themeContent) {
@@ -123,6 +131,7 @@ public class Oogiri {
 
     /**
      * お題作成日時を取得します。
+     * 
      * @return お題作成日時
      */
     public LocalDateTime getThemeCreatedAt() {
@@ -131,6 +140,7 @@ public class Oogiri {
 
     /**
      * お題作成日時を設定します。
+     * 
      * @param themeCreatedAt お題作成日時
      */
     public void setThemeCreatedAt(LocalDateTime themeCreatedAt) {
@@ -139,6 +149,7 @@ public class Oogiri {
 
     /**
      * お題更新日時を取得します。
+     * 
      * @return お題更新日時
      */
     public LocalDateTime getThemeUpdatedAt() {
@@ -147,6 +158,7 @@ public class Oogiri {
 
     /**
      * お題更新日時を設定します。
+     * 
      * @param themeUpdatedAt お題更新日時
      */
     public void setThemeUpdatedAt(LocalDateTime themeUpdatedAt) {
@@ -155,6 +167,7 @@ public class Oogiri {
 
     /**
      * 回答IDを取得します。
+     * 
      * @return 回答ID
      */
     public int getAnswerId() {
@@ -163,6 +176,7 @@ public class Oogiri {
 
     /**
      * 回答IDを設定します。
+     * 
      * @param answerId 回答ID
      */
     public void setAnswerId(int answerId) {
@@ -171,6 +185,7 @@ public class Oogiri {
 
     /**
      * 回答投稿ユーザーIDを取得します。
+     * 
      * @return 回答投稿ユーザーID
      */
     public int getAnswerUserId() {
@@ -179,6 +194,7 @@ public class Oogiri {
 
     /**
      * 回答投稿ユーザーIDを設定します。
+     * 
      * @param answerUserId 回答投稿ユーザーID
      */
     public void setAnswerUserId(int answerUserId) {
@@ -187,6 +203,7 @@ public class Oogiri {
 
     /**
      * 回答ユーザーネームを取得します。
+     * 
      * @return 回答ユーザーネーム
      */
     public String getAnswerUserName() {
@@ -195,6 +212,7 @@ public class Oogiri {
 
     /**
      * 回答ユーザーネームを設定します。
+     * 
      * @param answerUserName 回答ユーザーネーム
      */
     public void setAnswerUserName(String answerUserName) {
@@ -203,6 +221,7 @@ public class Oogiri {
 
     /**
      * 回答内容を取得します。
+     * 
      * @return 回答内容
      */
     public String getAnswerContent() {
@@ -211,6 +230,7 @@ public class Oogiri {
 
     /**
      * 回答内容を設定します。
+     * 
      * @param answerContent 回答内容
      */
     public void setAnswerContent(String answerContent) {
@@ -219,6 +239,7 @@ public class Oogiri {
 
     /**
      * 回答作成日時を取得します。
+     * 
      * @return 回答作成日時
      */
     public LocalDateTime getAnswerCreatedAt() {
@@ -227,6 +248,7 @@ public class Oogiri {
 
     /**
      * 回答作成日時を設定します。
+     * 
      * @param answerCreatedAt 回答作成日時
      */
     public void setAnswerCreatedAt(LocalDateTime answerCreatedAt) {
@@ -235,6 +257,7 @@ public class Oogiri {
 
     /**
      * 回答削除日時を取得します。
+     * 
      * @return 回答削除日時
      */
     public LocalDateTime getAnswerDeletedAt() {
@@ -243,6 +266,7 @@ public class Oogiri {
 
     /**
      * 回答削除日時を設定します。
+     * 
      * @param answerDeletedAt 回答削除日時
      */
     public void setAnswerDeletedAt(LocalDateTime answerDeletedAt) {
@@ -251,22 +275,25 @@ public class Oogiri {
 
     /**
      * 回答リストを取得します。
+     * 
      * @return 回答リスト
      */
-    public List<OogiriAnswerResponse> getAnswers() {
+    public List<OogiriAnswerResources> getAnswers() {
         return answers;
     }
 
     /**
      * 回答リストを設定します。
+     * 
      * @param answers 回答リスト
      */
-    public void setAnswers(List<OogiriAnswerResponse> answers) {
+    public void setAnswers(List<OogiriAnswerResources> answers) {
         this.answers = answers;
     }
 
     /**
      * リアクションIDを取得します。
+     * 
      * @return リアクションID
      */
     public int getReactionId() {
@@ -275,6 +302,7 @@ public class Oogiri {
 
     /**
      * リアクションIDを設定します。
+     * 
      * @param reactionId リアクションID
      */
     public void setReactionId(int reactionId) {
@@ -283,6 +311,7 @@ public class Oogiri {
 
     /**
      * リアクションIDリスト(文字列)を取得します。
+     * 
      * @return リアクションIDリスト(文字列)
      */
     public String getReactionIds() {
@@ -291,6 +320,7 @@ public class Oogiri {
 
     /**
      * リアクションIDリスト(文字列)を設定します。
+     * 
      * @param reactionIds リアクションIDリスト(文字列)
      */
     public void setReactionIds(String reactionIds) {
@@ -299,6 +329,7 @@ public class Oogiri {
 
     /**
      * リアクションユーザーIDを取得します。
+     * 
      * @return リアクションユーザーID
      */
     public int getReactionUserId() {
@@ -307,6 +338,7 @@ public class Oogiri {
 
     /**
      * リアクションユーザーIDを設定します。
+     * 
      * @param reactionUserId リアクションユーザーID
      */
     public void setReactionUserId(int reactionUserId) {
@@ -315,6 +347,7 @@ public class Oogiri {
 
     /**
      * リアクションユーザーIDリスト(文字列)を取得します。
+     * 
      * @return リアクションユーザーIDリスト(文字列)
      */
     public String getReactionUserIds() {
@@ -323,6 +356,7 @@ public class Oogiri {
 
     /**
      * リアクションユーザーIDリスト(文字列)を設定します。
+     * 
      * @param reactionUserIds リアクションユーザーIDリスト(文字列)
      */
     public void setReactionUserIds(String reactionUserIds) {
@@ -331,6 +365,7 @@ public class Oogiri {
 
     /**
      * リアクションステータスを取得します。
+     * 
      * @return リアクションステータス
      */
     public int getReactionStatus() {
@@ -339,6 +374,7 @@ public class Oogiri {
 
     /**
      * リアクションステータスを設定します。
+     * 
      * @param reactionStatus リアクションステータス
      */
     public void setReactionStatus(int reactionStatus) {
@@ -347,6 +383,7 @@ public class Oogiri {
 
     /**
      * リアクションステータスリスト（文字列）を取得します。
+     * 
      * @return リアクションステータスリスト（文字列）
      */
     public String getReactionStatuses() {
@@ -355,6 +392,7 @@ public class Oogiri {
 
     /**
      * リアクションステータスリスト（文字列）を設定します。
+     * 
      * @param reactionStatuses リアクションステータスリスト（文字列）
      */
     public void setReactionStatuses(String reactionStatuses) {
@@ -363,6 +401,7 @@ public class Oogiri {
 
     /**
      * リアクション作成日時を取得します。
+     * 
      * @return リアクション作成日時
      */
     public LocalDateTime getReactionCreatedAt() {
@@ -371,6 +410,7 @@ public class Oogiri {
 
     /**
      * リアクション作成日時を設定します。
+     * 
      * @param reactionCreatedAt リアクション作成日時
      */
     public void setReactionCreatedAt(LocalDateTime reactionCreatedAt) {
@@ -379,6 +419,7 @@ public class Oogiri {
 
     /**
      * リアクション更新日時を取得します。
+     * 
      * @return リアクション更新日時
      */
     public LocalDateTime getReactionUpdatedAt() {
@@ -387,6 +428,7 @@ public class Oogiri {
 
     /**
      * リアクション更新日時を設定します。
+     * 
      * @param reactionUpdatedAt リアクション更新日時
      */
     public void setReactionUpdatedAt(LocalDateTime reactionUpdatedAt) {
@@ -395,6 +437,7 @@ public class Oogiri {
 
     /**
      * リアクション数を取得します。
+     * 
      * @return リアクション数
      */
     public int getReactionNum() {
@@ -403,6 +446,7 @@ public class Oogiri {
 
     /**
      * リアクション数を設定します。
+     * 
      * @param reactionNum リアクション数
      */
     public void setReactionNum(int reactionNum) {
